communiquer/requirements.txt
Django==3.0.6
djangorestframework==3.10.3
django-heroku==0.3.1
psycopg2==2.8.3
psycopg2-binary==2.8.5
gunicorn==19.9.0

